refactor: Remove standard env
