+++ gym_simplifiedtetris/envs/__init__.py
@@ -1,7 +1,6 @@
 from .simplified_tetris_binary_env import SimplifiedTetrisBinaryEnv
 from .simplified_tetris_engine import SimplifiedTetrisEngine
 from .simplified_tetris_base_env import SimplifiedTetrisBaseEnv
-from .simplified_tetris_standard_env import SimplifiedTetrisStandardEnv
 from .simplified_tetris_part_binary_env import SimplifiedTetrisPartBinaryEnv
 from .reward_shaping import (
     SimplifiedTetrisBinaryShapedEnv,
@@ -12,7 +11,6 @@ __all__ = [
     "SimplifiedTetrisBinaryEnv",
     "SimplifiedTetrisEngine",
     "SimplifiedTetrisBaseEnv",
-    "SimplifiedTetrisStandardEnv",
     "SimplifiedTetrisPartBinaryEnv",
     "SimplifiedTetrisBinaryShapedEnv",
     "SimplifiedTetrisPartBinaryShapedEnv",
